Refactor: Move controller structure
