+++ pkg/webhook/operatingsystemconfig/image/add.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package operatingsystemconfig
+package image
 
 import (
 	extensionswebhook "github.com/gardener/gardener/extensions/pkg/webhook"
