+++ pkg/webhook/operatingsystemconfig/image/mutator.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package operatingsystemconfig
+package image
 
 import (
 	"context"
