+++ pkg/webhook/operatingsystemconfig/image/mutator_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package operatingsystemconfig_test
+package image_test
 
 import (
 	"context"
@@ -19,7 +19,7 @@ import (
 	fakeclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/gardener/gardener-extension-image-rewriter/pkg/apis/config/v1alpha1"
-	. "github.com/gardener/gardener-extension-image-rewriter/pkg/webhook/operatingsystemconfig"
+	. "github.com/gardener/gardener-extension-image-rewriter/pkg/webhook/operatingsystemconfig/image"
 )
 
 var _ = Describe("Mutator", func() {
